refactor: refactor data plane validationRule to the core data plane
