+++ core/data-plane/data-plane-util/src/main/java/org/eclipse/dataspaceconnector/dataplane/common/validation/EmptyValueValidationRule.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.aws.dataplane.s3.validation;
+package org.eclipse.dataspaceconnector.dataplane.common.validation;
 
 import org.eclipse.dataspaceconnector.common.string.StringUtils;
 import org.eclipse.dataspaceconnector.spi.result.Result;
