+++ core/data-plane/data-plane-util/src/test/java/org/eclipse/dataspaceconnector/dataplane/common/validation/ValidationRuleTest.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.aws.dataplane.s3.validation;
+package org.eclipse.dataspaceconnector.dataplane.common.validation;
 
 import org.eclipse.dataspaceconnector.spi.result.Result;
 import org.junit.jupiter.api.Test;
