+++ extensions/data-plane/data-plane-google-storage/src/test/java/org/eclipse/dataspaceconnector/gcp/dataplane/storage/GcsDataSourceFactoryTest.java
@@ -23,6 +23,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 
 class GcsDataSourceFactoryTest {
